renamed package to jcountdown
