+++ src/main/java/org/adoptopenjdk/jcountdown/RESTConfig.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.adoptopenjdk.javacountdown;
+package org.adoptopenjdk.jcountdown;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
