+++ src/main/java/org/adoptopenjdk/jcountdown/boundary/DataProvider.java
@@ -13,10 +13,10 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.adoptopenjdk.javacountdown.boundary;
+package org.adoptopenjdk.jcountdown.boundary;
 
-import org.adoptopenjdk.javacountdown.control.MongoDatastore;
-import org.adoptopenjdk.javacountdown.entity.*;
+import org.adoptopenjdk.jcountdown.control.MongoDatastore;
+import org.adoptopenjdk.jcountdown.entity.*;
 
 import javax.ejb.Stateless;
 import javax.enterprise.event.Event;
