+++ src/main/java/org/adoptopenjdk/jcountdown/boundary/VersionResource.java
@@ -13,9 +13,9 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.adoptopenjdk.javacountdown.boundary;
+package org.adoptopenjdk.jcountdown.boundary;
 
-import org.adoptopenjdk.javacountdown.entity.VisitTransfer;
+import org.adoptopenjdk.jcountdown.entity.VisitTransfer;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
