+++ src/main/java/org/adoptopenjdk/jcountdown/control/MongoClientProducer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.adoptopenjdk.javacountdown.control;
+package org.adoptopenjdk.jcountdown.control;
 
 import com.mongodb.MongoClient;
