+++ src/main/java/org/adoptopenjdk/jcountdown/entity/Visit.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.adoptopenjdk.javacountdown.entity;
+package org.adoptopenjdk.jcountdown.entity;
 
 
 import org.bson.types.ObjectId;
